# ecommerce-shopping-cart/backend/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "E-Commerce Shopping Cart API",
    "description": "E-Commerce Shopping Cart System API documentation",
    "version": "1.0.0",
    "contact": {
      "name": "API Support",
      "email": "support@ecommercesite.com"
    }
  },
  "host": "localhost:5000",
  "schemes": [
    "http"
  ],
  "basePath": "/api/v1",
...
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\""
    }
  },
  "paths": {
...
    "/cart": {
      "get": {
        "tags": [
          "Cart"
        ],
        "summary": "Get user's cart items",
        "security": [
          {
            "Bearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Cart items for the user",
            "schema": {
              "$ref": "#/definitions/CartResponse"
            }
          }
        }
      },
      "post": {
        "tags": [
          "Cart"
        ],
        "summary": "Add item to cart",
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Product and quantity to add to cart",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AddToCart"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Item added to cart successfully",
            "schema": {
              "$ref": "#/definitions/CartItemResponse"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Cart"
        ],
        "summary": "Clear user's cart",
        "security": [
          {
            "Bearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Cart cleared successfully"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Product": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "Laptop"
        },
        "description": {
          "type": "string",
          "example": "High-performance laptop"
        },
        "price": {
          "type": "number",
          "example": 999.99
        },
        "category": {
          "type": "string",
          "example": "Electronics"
        },
        "inventory": {
          "type": "integer",
          "example": 50
        },
        "images": {
          "type": "array",
          "items": {
            "type": "string",
            "example": "https://example.com/image.jpg"
          }
        }
      }
    },
...
    "AddToCart": {
      "type": "object",
      "properties": {
        "productId": {
          "type": "integer",
          "example": 1
        },
        "quantity": {
          "type": "integer",
          "example": 2
        }
      }
    },
    "CartItemResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean",
          "example": true
        },
        "data": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "example": 1
            },
            "userId": {
              "type": "integer",
              "example": 1
            },
            "productId": {
              "type": "integer",
              "example": 1
            },
            "quantity": {
              "type": "integer",
              "example": 2
            }
          }
        }
      }
    },
    "CartResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean",
          "example": true
        },
        "count": {
          "type": "integer",
          "example": 3
        },
        "total": {
          "type": "number",
          "example": 150.75
        },
        "data": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer",
                "example": 1
              },
              "quantity": {
                "type": "integer",
                "example": 2
              },
              "Product": {
                "$ref": "#/definitions/Product"
              }
            }
          }
        }
      }
    },
...
  }
}